+++ tfx/experimental/templates/taxi/e2e_tests/test_utils.py
@@ -19,8 +19,13 @@ from __future__ import division
 from __future__ import print_function
 
 import codecs
+import datetime
 import locale
 import os
+import subprocess
+import tarfile
+import time
+import urllib.request
 
 from typing import Text, List, Iterable, Tuple
 
@@ -28,9 +33,12 @@ from absl import logging
 from click import testing as click_testing
 import tensorflow as tf
 
+from tfx.orchestration.kubeflow import test_utils as kubeflow_test_utils
 from tfx.tools.cli.cli_main import cli_group
 from tfx.utils import io_utils
-
+from tfx.utils import telemetry_utils
+import yaml
+from google.cloud import storage
 
 class BaseEndToEndTest(tf.test.TestCase):
   """This test covers step 1~6 of the accompanying document[1] for taxi template.
@@ -62,7 +70,17 @@ class BaseEndToEndTest(tf.test.TestCase):
     super(BaseEndToEndTest, self).tearDown()
     os.chdir(self._old_cwd)
 
-  def _runCli(self, args: List[Text]) -> click_testing.Result:
+  def _cleanup_with_retry(self, method):
+    max_num_trial = 3
+    for _ in range(max_num_trial):
+      try:
+        method()
+      except Exception as err:  # pylint:disable=broad-except
+        logging.info(err)
+      else:
+        break
+
+  def _runCli(self, args: List[Text]) -> click_testing.Result:  # pylint:disable=invalid-name
     logging.info('Running cli: %s', args)
     result = self._cli_runner.invoke(cli_group, args)
     logging.info('%s', result.output)
@@ -73,7 +91,7 @@ class BaseEndToEndTest(tf.test.TestCase):
 
     return result
 
-  def _addAllComponents(self) -> Text:
+  def _addAllComponents(self) -> Text:  # pylint:disable=invalid-name
     """Change 'pipeline.py' file to put all components into the pipeline."""
     return self._uncomment(
         os.path.join('pipeline', 'pipeline.py'), ['components.append('])
@@ -83,7 +101,7 @@ class BaseEndToEndTest(tf.test.TestCase):
     replacements = [('# ' + s, s) for s in expressions]
     return self._replaceFileContent(filepath, replacements)
 
-  def _replaceFileContent(self, filepath: Text,
+  def _replaceFileContent(self, filepath: Text,  # pylint:disable=invalid-name
                           replacements: Iterable[Tuple[Text, Text]]) -> Text:
     """Update given file using `replacements`."""
     path = os.path.join(self._project_dir, filepath)
@@ -94,7 +112,7 @@ class BaseEndToEndTest(tf.test.TestCase):
     io_utils.write_string_file(path, content)
     return path
 
-  def _uncommentMultiLineVariables(self, filepath: Text,
+  def _uncommentMultiLineVariables(self, filepath: Text,  # pylint:disable=invalid-name
                                    variables: Iterable[Text]) -> Text:
     """Update given file by uncommenting a variable.
 
@@ -150,7 +168,7 @@ class BaseEndToEndTest(tf.test.TestCase):
     io_utils.write_string_file(path, ''.join(result))
     return path
 
-  def _copyTemplate(self):
+  def _copyTemplate(self):  # pylint:disable=invalid-name
     result = self._runCli([
         'template',
         'copy',
@@ -163,3 +181,198 @@ class BaseEndToEndTest(tf.test.TestCase):
     ])
     self.assertEqual(0, result.exit_code)
     self.assertIn('Copying taxi pipeline template', result.output)
+
+  def _get_kfp_runs(self):
+    # CLI uses experiment_name which is the same as pipeline_name.
+    experiment_id = self._kfp_client.get_experiment(
+        experiment_name=self._pipeline_name).id
+    response = self._kfp_client.list_runs(experiment_id=experiment_id)
+    return response.runs
+
+  def _delete_runs(self):
+    for run in self._get_kfp_runs():
+      self._kfp_client._run_api.delete_run(id=run.id)  # pylint: disable=protected-access
+
+  def _delete_pipeline(self):
+    self._runCli([
+        'pipeline', 'delete', '--engine', 'kubeflow', '--pipeline_name',
+        self._pipeline_name, '--endpoint', self._endpoint
+    ])
+
+  def _delete_pipeline_data(self):
+    path = 'tfx_pipeline_output/{}'.format(self._pipeline_name)
+    orchestration_test_utils.delete_gcs_files(self._GCP_PROJECT_ID,
+                                              self._BUCKET_NAME, path)
+    path = '{}/{}'.format(self._DATA_DIRECTORY_NAME, self._pipeline_name)
+    orchestration_test_utils.delete_gcs_files(self._GCP_PROJECT_ID,
+                                              self._BUCKET_NAME, path)
+
+  def _delete_base_container_image(self):
+    subprocess.check_output([
+        'gcloud', 'container', 'images', 'delete', self._base_container_image
+    ])
+
+  def _delete_target_container_image(self):
+    subprocess.check_output([
+        'gcloud', 'container', 'images', 'delete', self._target_container_image
+    ])
+
+  def _prepare_data(self):
+    client = storage.Client(project=self._GCP_PROJECT_ID)
+    bucket = client.bucket(self._BUCKET_NAME)
+    blob = bucket.blob('{}/{}/data.csv'.format(self._DATA_DIRECTORY_NAME,
+                                               self._pipeline_name))
+    blob.upload_from_filename('data/data.csv')
+
+  def _get_endpoint(self):  # pylint: disable=inconsistent-return-statements
+    output = subprocess.check_output(
+        'kubectl describe configmap inverse-proxy-config -n kubeflow'.split())
+    for line in output.decode('utf-8').split('\n'):
+      if line.endswith('googleusercontent.com'):
+        return line
+
+  def _prepare_skaffold(self):
+    self._skaffold = os.path.join(self._temp_dir, 'skaffold')
+    urllib.request.urlretrieve(
+        'https://storage.googleapis.com/skaffold/releases/latest/skaffold-linux-amd64',
+        self._skaffold)
+    os.chmod(self._skaffold, 0o775)
+
+  def _create_pipeline(self):
+    result = self._runCli([
+        'pipeline',
+        'create',
+        '--engine',
+        'kubeflow',
+        '--pipeline_path',
+        'kubeflow_dag_runner.py',
+        '--endpoint',
+        self._endpoint,
+        '--build-target-image',
+        self._target_container_image,
+        '--skaffold-cmd',
+        self._skaffold,
+        '--build-base-image',
+        self._base_container_image,
+    ])
+    self.assertEqual(0, result.exit_code)
+
+  def _update_pipeline(self):
+    result = self._runCli([
+        'pipeline',
+        'update',
+        '--engine',
+        'kubeflow',
+        '--pipeline_path',
+        'kubeflow_dag_runner.py',
+        '--endpoint',
+        self._endpoint,
+        '--skaffold-cmd',
+        self._skaffold,
+    ])
+    self.assertEqual(0, result.exit_code)
+
+  def _run_pipeline(self):
+    result = self._runCli([
+        'run',
+        'create',
+        '--engine',
+        'kubeflow',
+        '--pipeline_name',
+        self._pipeline_name,
+        '--endpoint',
+        self._endpoint,
+    ])
+    self.assertEqual(0, result.exit_code)
+    self._wait_until_completed(self._parse_run_id(result.output))
+
+  def _parse_run_id(self, output: str):
+    run_id_lines = [
+        line for line in output.split('\n')
+        if '| {} |'.format(self._pipeline_name) in line
+    ]
+    self.assertLen(run_id_lines, 1)
+    return run_id_lines[0].split('|')[2].strip()
+
+  def _wait_until_completed(self, run_id: str):
+    # This timeout will never expire. polling_count * interval == 20min.
+    timeout = datetime.timedelta(hours=1)
+    end_state = kubeflow_test_utils.poll_kfp_with_retry(
+        self._endpoint, run_id, self._MAX_POLLING_COUNT, timeout,
+        self._POLLING_INTERVAL_IN_SECONDS)
+    self.assertEqual(end_state.lower(), kubeflow_test_utils.KFP_SUCCESS_STATUS)
+
+  def _check_telemetry_label(self):
+    file_path = os.path.join(self._project_dir,
+                             '{}.tar.gz'.format(self._pipeline_name))
+    self.assertTrue(tf.io.gfile.exists(file_path))
+
+    with tarfile.TarFile.open(file_path).extractfile(
+        'pipeline.yaml') as pipeline_file:
+      self.assertIsNotNone(pipeline_file)
+      pipeline = yaml.safe_load(pipeline_file)
+      metadata = [
+          c['metadata'] for c in pipeline['spec']['templates'] if 'dag' not in c
+      ]
+      for m in metadata:
+        self.assertEqual('tfx-template',
+                         m['labels'][telemetry_utils.LABEL_KFP_SDK_ENV])
+
+  def _get_grpc_port(self) -> Text:
+    """Get the port number used by MLMD gRPC server."""
+    get_grpc_port_command = [
+        'kubectl', '-n', 'kubeflow', 'get', 'configmap',
+        'metadata-grpc-configmap', '-o',
+        'jsonpath={.data.METADATA_GRPC_SERVICE_PORT}'
+    ]
+
+    grpc_port = subprocess.check_output(get_grpc_port_command)
+    return grpc_port.decode('utf-8')
+
+  def _setup_mlmd_port_forward(self) -> subprocess.Popen:
+    """Uses port forward to talk to MLMD gRPC server."""
+    grpc_port = self._get_grpc_port()
+
+    is_bind = False
+
+    for port in range(self._KFP_E2E_TEST_FORWARDING_PORT_BEGIN,
+                      self._KFP_E2E_TEST_FORWARDING_PORT_END):
+      grpc_forward_command = [
+          'kubectl', 'port-forward', 'deployment/metadata-grpc-deployment',
+          '-n', 'kubeflow', ('%s:%s' % (port, grpc_port))
+      ]
+      # Begin port forwarding.
+      proc = subprocess.Popen(grpc_forward_command)
+      try:
+        # Wait while port forward to pod is being established
+        poll_grpc_port_command = ['lsof', '-i', ':%s' % port]
+        result = subprocess.run(  # pylint: disable=subprocess-run-check
+            poll_grpc_port_command,
+            stdout=subprocess.PIPE)
+        for _ in range(self._MAX_ATTEMPTS):
+          if (result.returncode == 0 and
+              'kubectl' in result.stdout.decode('utf-8')):
+            is_bind = True
+            break
+          logging.info(
+              'Waiting while gRPC port-forward is being established...')
+          time.sleep(5)
+          result = subprocess.run(  # pylint: disable=subprocess-run-check
+              poll_grpc_port_command,
+              stdout=subprocess.PIPE)
+
+      except:  # pylint: disable=bare-except
+        # Kill the process in case unexpected error occurred.
+        proc.kill()
+
+      if is_bind:
+        self._port = port
+        break
+
+    if not is_bind:
+      raise RuntimeError('Failed to establish gRPC port-forward to cluster in '
+                         'the specified range: port %s to %s' %
+                         (self._KFP_E2E_TEST_FORWARDING_PORT_BEGIN,
+                          self._KFP_E2E_TEST_FORWARDING_PORT_END))
+
+    return proc
